tools/tui/pyproject.toml
[project]
name = "relais-tui"
version = "0.1.0"
description = "Terminal User Interface client for RELAIS via REST SSE API"
requires-python = ">=3.11"
dependencies = [
    "textual>=3.0,<4.0",
    "httpx>=0.28,<1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.scripts]
relais-tui = "relais_tui.__main__:main"
relais = "relais_tui.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/relais_tui"]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3",
    "pytest-cov>=7.1.0",
    "textual-dev>=1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
